Ported /permissions to permissions.go
